Implemented PLACE request and minimal server
This commit contains the implementation of a bunch of functions so the game is
able to send its PLACE requests. Besides, a temporary server application has
been created. It does nothing particular, but is able to receive requests and
print them, for testing purposes.

Adresses are hardcoded to the loopback address (127.0.0.1) and port 5000.