Firing implemented
After having placed all his/her boats, the player can now chose a square to
fire upon. The client application will send a `FIRE` request if the given
coordinates are consistent. The application will then wait for the server to
answer with a `HIT`or a `MISS` request and then update its tracking grid
accordingly.

However, since the application only waits for one response, it will not catch
any `SINK` request by now. This improvement is to be done later.