Overall-network: WIP #1
Now, a client can create a listening socket, send its listening socket infos to
the server through a INFO request, ask for available games with a GAMES request,
receive available games thanks to several GAME requests sent by the server.

It can also connect to another player's listening socket and send the player's
name to the opponent with a NAME request.

More to come, last features of the network layer and code cleaning.

Integration of the network layer is likely to be done in another branch, later.