Reworked in-game network layer
The in-game network layer has been update to match the new network structure.
Plus, the in-game workflow has been simplified and made more easily maintainable.

However, I couldn't test anything since the `game` test module is unable to
establish a dialog socket between two clients by itself. So, it may be buggy and
some fixes are likely to be necessary.